Refactor to reduce main.go size
